ai-service/requirements.txt
python-dotenv==1.0.0
pytest==8.3.4
streamlit==1.41.1
watchdog==6.0.0
langchain-openai==0.3.1
ratelimit